Section, lane, field and movement modifications
Made a lane class to better categorize each field properly and support the new movement logic.
Modified the Field class by adding IsPossible bool to support the movement logic
Implemented movement logic onto the racerUnit script. The logic works, but needs to be cleaned and some of the logic will be moved elsewhere once the turns are more properly managed