Corners and sectionshifts
Implemented logic into the racer unit code to shift to the next section when at the end of a section.
Implemented logic to check the lane shift rule.
Added a field to Section scripts so a section is aware of which section comes next.
added a corner to my scene and modified it to work with the movement logic I implemented